Géolocalisation, MAJ Migration * Automatique si possible. * Afficher une input d'adresse sinon. * Changer le stoackage de la position de float vers double(15,8) pour une plus grande précision.

+++ app/database/migrations/2015_01_10_205813_creer_table_utilisateurs.php
@@ -43,8 +43,8 @@ class CreerTableUtilisateurs extends Migration {
 
 			$table->string('titre'); // ->unique(); (unique ou non?)
 			$table->text('rapport');	
-			$table->float('geo-x');
-			$table->float('geo-y');
+			$table->double('geo-x', 15, 8);
+			$table->double('geo-y', 15, 8);
 			$table->string('adresse');
 
 			$table->boolean('afficher');
